+++ CHANGELOG.md
@@ -5,7 +5,6 @@
 * support arm deployments
 
 #### 4.1.0 - TODO ITEMS
-* run script every time the node starts (crontab @reboot)
 * add option to deploy AMS (monitoring stack)
 * add install script to install basic linux tools (ip-utils, dns-utils, ping, telnet, netcat, netstat, ss, iostat, mpstat, sar)
 
@@ -22,6 +21,7 @@
 * update readme and help pages
 * update scripts/
 * make help pages work without verifying if backend is working
+* install a script to run before aerospike starts and after it stops
 
 #### v3.1.2
 * `make-cluster` and `cluster-grow`: automatically add `cluster-name` to `aerospike.conf` unless specified not to by the `-O` switch